+++ src/protobuf-net.Core/Serializers/MapSerializer.External.cs
@@ -9,7 +9,7 @@ namespace ProtoBuf.Serializers
     /// ExternalMapSerializer provides a base class for concrete types to inherit from, but only provide the methods for collection management
     /// It does not require changes to internal protobuf-net state handling
     /// </summary>
-    /// <typeparam name="TCollection">the collection type being provided (e.g. Map<_,_> for F#) </typeparam>
+    /// <typeparam name="TCollection">the collection type being provided (e.g. Map for F#) </typeparam>
     /// <typeparam name="TKey">key to the collection</typeparam>
     /// <typeparam name="TValue">type of the value held within the collection</typeparam>
     public abstract class ExternalMapSerializer<[DynamicallyAccessedMembers(DynamicAccess.ContractType)]TCollection, [DynamicallyAccessedMembers(DynamicAccess.ContractType)] TKey, [DynamicallyAccessedMembers(DynamicAccess.ContractType)] TValue> : MapSerializer<TCollection, TKey, TValue> where TCollection : IEnumerable<KeyValuePair<TKey, TValue>>