+++ docs/releasenotes.md
@@ -13,7 +13,7 @@ Packages are available on NuGet: [protobuf-net](https://www.nuget.org/packages/p
 
 ## pending
 
-- support unknwon/extension fields on models that involve inheritance
+- support unknwon/extension fields on models that involve inheritance (via either `Extensible` or `ITypedExtensible`)
 
 ## 3.0.62